requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
Authlib==1.2.0
autopep8==2.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
exceptiongroup==1.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy-extensions==1.0.0
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
rfc3986==1.5.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tda-api==1.6.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
websockets==10.4
wsproto==1.2.0
